Rename internal headers to conform with + naming convention
