+++ AsyncDisplayKit/Private/ASDisplayNode+FrameworkPrivate.h
@@ -12,7 +12,7 @@
 //
 
 #import "_AS-objc-internal.h"
-#import "ASDisplayNodeExtraIvars.h"
+#import "ASDisplayNode+ExtraIvars.h"
 #import "ASDisplayNode.h"
 #import "ASSentinel.h"
 #import "ASThread.h"
