+++ AsyncDisplayKit/Private/ASDisplayNode+UIViewBridge.mm
@@ -10,8 +10,8 @@
 #import "_ASPendingState.h"
 #import "ASInternalHelpers.h"
 #import "ASAssert.h"
-#import "ASDisplayNodeInternal.h"
-#import "ASDisplayNodeExtras.h"
+#import "ASDisplayNode+Internal.h"
+#import "ASDisplayNode+Extras.h"
 #import "ASDisplayNode+Subclasses.h"
 #import "ASDisplayNode+FrameworkPrivate.h"
 #import "ASDisplayNode+Beta.h"
